refactor(streams): move local-terminal PTY core behind EventSink
