+++ crates/streams/src/lib.rs
@@ -7,6 +7,7 @@ pub mod forward;
 pub mod helm;
 pub mod logs;
 pub mod sink;
+pub mod terminal;
 pub mod test_util;
 pub mod watch;
